refactor: Agreement 클래스 분리
